disentangle parsing code + update to C++17
